2to1 Shader

+++ Can/Can/src/Can/Camera/Camera.cpp
@@ -21,8 +21,8 @@ namespace Can::Camera
 	void Perspective::Turn()
 	{
 		float radius = 10.0f;
-		float camX = sin(glfwGetTime()) * radius;
-		float camZ = cos(glfwGetTime()) * radius;
+		float camX = (float)sin(glfwGetTime()) * radius;
+		float camZ = (float)cos(glfwGetTime()) * radius;
 		view = glm::lookAt(glm::vec3(camX, 0.0, camZ), glm::vec3(0.0, 0.0, 0.0), glm::vec3(0.0, 1.0, 0.0));
 	}
 	void Camera::RecalculateViewMatrix()
